Add teamable_helper, leagueable_helper, seasonable_helper and their test files. Framework added
