+++ lib/stat_tracker.rb
@@ -6,12 +6,19 @@ require_relative 'gameable'
 require_relative 'leagueable'
 require_relative 'teamable'
 require_relative 'seasonable'
+require_relative 'leagueable_helper'
+require_relative 'teamable_helper'
+require_relative 'seasonable_helper'
+
 
 class StatTracker
   include Gameable
   include Leagueable
   include Teamable
   include Seasonable
+  include LeagueableHelper
+  include TeamableHelper
+  include SeasonableHelper
 
   attr_reader :games, :teams, :game_teams
